+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 - Fixed #891, synchronizing re-evaluation, blur, and keyboard idle
 - Fixed #910, excluding custom emojis from character's emoji picker.
 - Fixed #895, resolving tooltip flicker.
+- Fixed #901, improved hover feedback on editable token text.
 
 ## Maintenance